+++ CHANGELOG.md
@@ -1,12 +1,16 @@
+## 2021-10-28
+*Improvements*
+- SDK update to 2.1.0.
+
 ## 2021-05-24
 *Fixes*
 - added to the input "URL" field. Implemented logic about using it as the start link;
-- removed logging of unnecessary and confusing information;
+- removed logging of unnecessary and confusing information.
 
 ## 2021-05-24
 *Fixes*
 - removed "company search" mode;
-- added additional values to the final output object;
+- added additional values to the final output object.
 
 ## 2021-05-13
 *Fixes*